refactor move all model interfaces to interfaces folder
