+++ src/interfaces/IAlbum.ts
@@ -1,0 +1,16 @@
+interface IAlbum {
+  id: string,
+  title: string,
+  artist: string,
+  trackCount: number,
+  imageUrl: string,
+  posterId: string,
+  pickedTracks: {
+    title: string,
+    trackNumber: number,
+    pickerIds: string[]
+  }[],
+  topTrackNumber: number
+}
+
+export default IAlbum;
