+++ src/interfaces/IMember.ts
@@ -1,0 +1,10 @@
+interface IMember {
+  id: string,
+  firstName: string,
+  lastName: string,
+  color: string,
+  participatedRoundIds: string[],
+  postedAlbumIds: string[],
+}
+
+export default IMember;
