+++ src/interfaces/IRound.ts
@@ -1,0 +1,12 @@
+interface IRound {
+  id: string,
+  number: number,
+  description: string,
+  participantIds: string[],
+  albumIds: string[],
+  startDate: string,
+  endDate: string,
+  picksPerParticipant: number
+}
+
+export default IRound;
